Initial project configuration
Set up project with Espressif example code, configure CMakeLists to
initialise ADF and set the board to lyrat V4.3.

Configure .clang-format and .editorconfig to help with consistent
formatting, these might have to be adjusted when we decide on a code
style.

Add .clangd to properly configure clangd lsp, in case this causes issues
for people using the vscode installation please let me know and I will
remove it from the main repository.